fix: WHOOP returns recovery_score as JSON number, not int
Both webhooks since Adam's strap went on were failing in the drainer
with JsonException. Path: $.records[0].score.recovery_score. Root
cause: WHOOP's v2 wire format serializes integer-valued fields like
recovery_score (46.0) and resting_heart_rate (71.0) as JSON numbers
with a trailing .0, but WhoopRecoveryScore declared them as int? and
System.Text.Json refused the coercion.

- WhoopRecoveryScore: RecoveryScore + RestingHeartRate now decimal?
  on the wire. The schema's int columns are still right — these
  values are integer-valued, just transmitted as numbers.

- RecoveryIngestionHandler rounds decimal? -> int? at the persistence
  boundary (MidpointRounding.AwayFromZero). NoOp comparison uses the
  rounded value so updates don't churn the row across re-fetches.

- WhoopApiClient.ParseAsync wraps deserialization in try/catch and
  logs the first 500 chars of the actual response body on
  JsonException. Caller gets Result.Failure with the JSON path so
  the next mismatch is debuggable in seconds, not from the database.

- Tests were already passing implicit int->decimal conversions in
  positional record ctor calls; no test changes needed.
- 21/21 tests still green.

Schema spec note: docs/schema/0001_initial_schema.sql claims "WHOOP
returns int" for recovery_score and resting_heart_rate. Empirically,
WHOOP returns numbers. The spec comment is wrong; updating in a
follow-up doc commit.

+++ src/SomaCore.Infrastructure/Whoop/WhoopModels.cs
@@ -25,11 +25,17 @@ public sealed record WhoopWebhookEnvelope(
     [property: JsonPropertyName("type")]     string EventType, // e.g. "recovery.updated"
     [property: JsonPropertyName("trace_id")] string TraceId);
 
-/// <summary>WHOOP v2 score sub-object on recovery.</summary>
+/// <summary>
+/// WHOOP v2 score sub-object on recovery. NB: WHOOP serializes <c>recovery_score</c>
+/// and <c>resting_heart_rate</c> as JSON numbers with a trailing <c>.0</c> even though
+/// the values are integer-valued. We type them as <c>decimal?</c> here so System.Text.Json
+/// accepts the wire format, and round to int at the persistence boundary in
+/// <see cref="SomaCore.Infrastructure.Recovery.RecoveryIngestionHandler"/>.
+/// </summary>
 public sealed record WhoopRecoveryScore(
-    [property: JsonPropertyName("recovery_score")]      int? RecoveryScore,
+    [property: JsonPropertyName("recovery_score")]      decimal? RecoveryScore,
     [property: JsonPropertyName("hrv_rmssd_milli")]     decimal? HrvRmssdMilli,
-    [property: JsonPropertyName("resting_heart_rate")]  int? RestingHeartRate,
+    [property: JsonPropertyName("resting_heart_rate")]  decimal? RestingHeartRate,
     [property: JsonPropertyName("spo2_percentage")]     decimal? Spo2Percentage,
     [property: JsonPropertyName("skin_temp_celsius")]   decimal? SkinTempCelsius,
     [property: JsonPropertyName("user_calibrating")]    bool? UserCalibrating);
